Upgrade linting config for new import-x plugin
