+++ apps/storybook/package.json
@@ -27,7 +27,7 @@
     "three": "0.185.1"
   },
   "devDependencies": {
-    "@esrf/eslint-config": "2.3.1",
+    "@esrf/eslint-config": "3.1.0",
     "@storybook/addon-docs": "10.4.6",
     "@storybook/addon-links": "10.4.6",
     "@storybook/react-vite": "10.4.6",
