+++ apps/storybook/src/ColorBar.stories.tsx
@@ -1,5 +1,5 @@
 import { ColorBar, type ColorBarProps, ScaleType } from '@h5web/lib';
-import { COLOR_SCALE_TYPES } from '@h5web/shared/vis-utils';
+import { COLOR_SCALE_TYPES } from '@h5web/shared/vis-models';
 import { type Meta, type StoryFn, type StoryObj } from '@storybook/react-vite';
 
 import FillHeight from './decorators/FillHeight';
