+++ apps/storybook/src/DomainWidget.stories.tsx
@@ -1,5 +1,5 @@
 import { type CustomDomain, DomainWidget, ScaleType } from '@h5web/lib';
-import { COLOR_SCALE_TYPES } from '@h5web/shared/vis-utils';
+import { COLOR_SCALE_TYPES } from '@h5web/shared/vis-models';
 import { type Meta, type StoryObj } from '@storybook/react-vite';
 import { useState } from 'react';
