+++ apps/storybook/src/HeatmapMesh.stories.tsx
@@ -7,8 +7,8 @@ import {
   VisCanvas,
 } from '@h5web/lib';
 import { assertDefined } from '@h5web/shared/guards';
-import { ScaleType } from '@h5web/shared/vis-models';
-import { COLOR_SCALE_TYPES, toTypedNdArray } from '@h5web/shared/vis-utils';
+import { COLOR_SCALE_TYPES, ScaleType } from '@h5web/shared/vis-models';
+import { toTypedNdArray } from '@h5web/shared/vis-utils';
 import { range } from 'd3-array';
 import ndarray from 'ndarray';
 import { LinearFilter, NearestFilter } from 'three';
