+++ apps/storybook/src/TiledHeatmapMesh/TiledHeatmapMesh.stories.tsx
@@ -7,8 +7,7 @@ import {
   TiledTooltipMesh,
   VisCanvas,
 } from '@h5web/lib';
-import { ScaleType } from '@h5web/shared/vis-models';
-import { COLOR_SCALE_TYPES } from '@h5web/shared/vis-utils';
+import { COLOR_SCALE_TYPES, ScaleType } from '@h5web/shared/vis-models';
 import { type Meta, type StoryObj } from '@storybook/react-vite';
 
 import FillHeight from '../decorators/FillHeight';
