+++ package.json
@@ -40,7 +40,7 @@
     "postversion": "git push && git push --tags"
   },
   "devDependencies": {
-    "@esrf/eslint-config": "2.3.1",
+    "@esrf/eslint-config": "3.1.0",
     "@simonsmith/cypress-image-snapshot": "10.0.4",
     "@testing-library/cypress": "10.1.3",
     "@types/node": "^24.13.2",
@@ -53,7 +53,6 @@
     "prettier": "3.9.4",
     "typescript": "5.9.3",
     "vitest": "4.1.9",
-    "vitest-fail-on-console": "0.10.1",
     "wait-on": "9.0.10"
   }
 }
