+++ packages/app/src/vis-packs/core/surface/SurfaceToolbar.tsx
@@ -7,7 +7,7 @@ import {
   SnapshotBtn,
   Toolbar,
 } from '@h5web/lib';
-import { COLOR_SCALE_TYPES } from '@h5web/shared/vis-utils';
+import { COLOR_SCALE_TYPES } from '@h5web/shared/vis-models';
 
 import { type SurfaceConfig } from './config';
