+++ packages/h5wasm/package.json
@@ -52,7 +52,7 @@
     "nanoid": "5.1.16"
   },
   "devDependencies": {
-    "@esrf/eslint-config": "2.3.1",
+    "@esrf/eslint-config": "3.1.0",
     "@h5web/app": "workspace:*",
     "@h5web/shared": "workspace:*",
     "@rollup/plugin-alias": "6.0.0",
