+++ packages/lib/package.json
@@ -86,7 +86,7 @@
     "zustand": "5.0.14"
   },
   "devDependencies": {
-    "@esrf/eslint-config": "2.3.1",
+    "@esrf/eslint-config": "3.1.0",
     "@h5web/shared": "workspace:*",
     "@react-three/fiber": "8.18.0",
     "@rollup/plugin-alias": "6.0.0",
